requirements.txt
llama-cpp-python>=0.2.80
numpy>=1.26
pillow>=10.4
pyyaml>=6.0
open3d>=0.17
opencv-python>=4.10
rich>=13.9
dashscope>=1.13.0
openai>=1.0.0